LST now functions in python 3.8

+++ models/LST.py
@@ -16,7 +16,7 @@ class Model(nn.Module):
         self.hidS = skip;
         self.Ck = args.CNN_kernel;
         self.skip = args.skip;
-        self.pt = (self.P - self.Ck)/self.skip
+        self.pt = int((self.P - self.Ck)/self.skip)
         self.hw = args.highway_window
         self.conv1 = nn.Conv2d(1, self.hidC, kernel_size = (self.Ck, self.m));
         self.GRU1 = nn.GRU(self.hidC, self.hidR);
